+++ ricaun.Nuke.PackageBuilder/Components/IHazPackageBuilderProject.cs
@@ -29,6 +29,18 @@ namespace ricaun.Nuke.Components
         bool ReleaseBundle => TryGetValue<bool?>(() => ReleaseBundle) ?? false;
 
         /// <summary>
+        /// Add ProjectNameFolder on the Contents (default: false)
+        /// </summary>
+        [Parameter]
+        bool ProjectNameFolder => TryGetValue<bool?>(() => ProjectNameFolder) ?? false;
+
+        /// <summary>
+        /// Add ProjectVersionFolder on the Contents (default: false)
+        /// </summary>
+        [Parameter]
+        bool ProjectVersionFolder => TryGetValue<bool?>(() => ProjectVersionFolder) ?? false;
+
+        /// <summary>
         /// GetPackageBuilderProject
         /// </summary>
         /// <returns></returns>